Move app into component folder
